requirements.txt
asgiref==3.8.1
astroid==3.1.0
attrs==23.2.0
bandit==1.7.8
black==24.4.0
certifi==2024.2.2
click==8.1.7
cloudinary==1.40.0
colorama==0.4.6
dill==0.3.8
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-extra-fields==3.7.0
drf-yasg==1.21.7
filetype==1.2.0
gitdb==4.0.11
GitPython==3.1.40
inflection==0.5.1
isort==5.13.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pbr==6.0.0
pillow==10.3.0
platformdirs==4.2.0
Pygments==2.17.2
PyJWT==2.8.0
pylint==3.1.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
rich==13.7.1
rpds-py==0.18.0
six==1.16.0
smmap==5.0.1
sqlparse==0.5.0
stevedore==5.2.0
swagger-spec-validator==3.0.3
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
